% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heptathlon_points.R
\docType{data}
\name{heptathlon_points}
\alias{heptathlon_points}
\title{Heptathlon points data.}
\source{
The data were taken from Wikipedia at
\url{https://en.wikipedia.org/wiki/Athletics_at_the_2012_Summer_Olympics_\%E2\%80\%93_Women\%27s_heptathlon}.
}
\usage{
data(heptathlon_points)
}
\value{
A data frame with 29 rows and 1 variable in the column Points.
}
\description{
Final scores for the heptathlon at the 2012 Olympic Games in London for the 29 athletes
who completed all events and were not disqualified.
}
\examples{
data(heptathlon_points)
head(heptathlon_points)
}
\keyword{datasets}
